Make agent's speed smoothly drop to zero when approaching equilibrium
To facilitate mathematical analysis, the equations need to be made smooth near the equilibrium points. Prior to this change, the agent's speed was constant. (I had intended that the speed would instantaneously drop to zero if the agent was exactly at an equilibrium point, but I now realize that in fact it would have evaluated to NaN instead when normalizing the zero-length perceived direction vector.) This change ensures that the agent's speed smoothly reduces to zero as it approaches a stable equilibrium point, while still limiting the maximum speed. Instead of calculating a direction of travel only (Agent.ComputeDirection), a velocity is now calculated (Agent.ComputeVelocity), and an explicit time step parameter was added to Agent.