Use topographic data that comes built into MATLAB
topo60c.mat was buried in a MATLAB example, so it wasn't accessible on the default file search path, and it was unclear whether committing it directly to this repo would violate licensing. This alternative topo.mat file is accessible from the default path, so it requires nothing to be committed, and it contains essentially the same data.